requirements.txt
alphabet_detector==0.0.7
beautifulsoup4
bleach==5.0.1
cachetools==4.2.2
CurrencyConverter
emoji==1.7.0
feedparser==6.0.10
future
geopy
markdown2
google_trans_new
humanize
jikanpy==4.3.2
lxml>=4.6.5
nltk>=3.6.6n
Pillow>=4.2.0
pretty_errors==1.2.25
Psutil==5.9.1
psycopg2-binary
pyrate-limiter==2.8.4
python-telegram-bot==13.14
regex==2022.7.25
requests
speedtest-cli
spongemock==0.3.4
sqlalchemy==1.4.46
telethon==1.24.0
wikipedia==1.4.0
zalgo-text==0.6
fontTools==4.28.5
pyrogram==2.0.41
telegraph==2.2.0
tgCrypto==1.2.5
youtube_dl
youtube_search==2.1.0
pytube==12.1.2
yt-dlp
youtube-search-python
pymongo
motor
dnspython==2.2.1
redis
gtts==2.3.0
cloudscraper==1.2.67
wget==3.2
tswift==0.7.0
lyricsgenius==3.0.1
envparse==0.2.0
cryptg~=0.4.0
cryptography
python-arq
aioredis==2.0.0
odmantic==0.3.5
aiohttp
hmtai
pykeyboard==0.1.5
aiofiles==0.8.0
bing_image_downloader==1.1.2
search_engine_parser
pytz==2022.6
html2text==2020.1.16
NumPy
opencv-python-headless==4.7.0.68
uvloop==0.17.0
python-dotenv==0.21.0
#googletrans==3.1.0a0
loguru
gpytranslate==1.4.0
apscheduler
tracemoepy
html5lib
lottie
natsort
ffmpeg-python
aiohttp[speedups]>=3.7.3
Ujson
git-python
multicolorcaptcha~=1.2.0
webcolors